Refactor use case package name from use_case to usecase
