+++ app/src/main/java/com/dogactnrvrdi/notesapp/domain/usecase/AddNoteUseCase.kt
@@ -1,4 +1,4 @@
-package com.dogactnrvrdi.notesapp.domain.use_case
+package com.dogactnrvrdi.notesapp.domain.usecase
 
 import com.dogactnrvrdi.notesapp.R
 import com.dogactnrvrdi.notesapp.data.model.InvalidNoteException
