+++ app/src/main/java/com/dogactnrvrdi/notesapp/domain/usecase/DeleteNoteUseCase.kt
@@ -1,12 +1,12 @@
-package com.dogactnrvrdi.notesapp.domain.use_case
+package com.dogactnrvrdi.notesapp.domain.usecase
 
 import com.dogactnrvrdi.notesapp.data.model.Note
 import com.dogactnrvrdi.notesapp.domain.repo.NoteRepository
 import javax.inject.Inject
 
-class GetNoteUseCase @Inject constructor(
+class DeleteNoteUseCase @Inject constructor(
     private val repo: NoteRepository
 ) {
-    suspend operator fun invoke(id: Int): Note? =
-        repo.getNoteById(id)
+    suspend operator fun invoke(note: Note) =
+        repo.deleteNote(note)
 }
