+++ app/src/main/java/com/dogactnrvrdi/notesapp/domain/usecase/GetNoteUseCase.kt
@@ -1,12 +1,12 @@
-package com.dogactnrvrdi.notesapp.domain.use_case
+package com.dogactnrvrdi.notesapp.domain.usecase
 
 import com.dogactnrvrdi.notesapp.data.model.Note
 import com.dogactnrvrdi.notesapp.domain.repo.NoteRepository
 import javax.inject.Inject
 
-class DeleteNoteUseCase @Inject constructor(
+class GetNoteUseCase @Inject constructor(
     private val repo: NoteRepository
 ) {
-    suspend operator fun invoke(note: Note) =
-        repo.deleteNote(note)
+    suspend operator fun invoke(id: Int): Note? =
+        repo.getNoteById(id)
 }
